requirements.txt
nbstripout

torch>=2.0.0
h5py
einops
tqdm
matplotlib
tensorboard
torch-tb-profiler
ninja
scipy
ipykernel
jupyter>=1.1.1
nbformat>=4.2.0
ipywidgets>=8.1.5
pydantic
aim
rich
lmdb

sentencepiece
pyav
torch_optimizer
pytorch-ignite
opencv-python
kornia
Pillow

optimum-quanto
accelerate
transformers
diffusers
peft
datasets
timm
bitsandbytes